Combine depthai simulate spatial detection into one file
